Initial commit with supperclass Fitter
