+++ Fitter/partiallyMerged/makeFits.py
@@ -1,3 +1,5 @@
+#!/usr/bin/env python
+
 import argparse
 import ROOT
 import sys
@@ -1034,11 +1036,6 @@ if __name__ == '__main__':
 
     fitter = WTaggingFitter(options)
 
-
-    if (not options.doWS): 
-        fitter.OpenWorkspace(options)
-
-    fitter.OpenWorkspace(options)
 
 #    if options.fitTT:
 #        print "Doing fits to matched tt MC. Tree must contain branch with flag for match/no-match to generator level W!"
